scripts/requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
requests==2.31.0
